[ML] Move property implementation to separate file.
